rename: notice 컴포넌트 auction-formd 으로 이동
